# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic SDLC Platform API
  description: REST API for the multi-agent SDLC orchestration platform.
  version: 0.4.0
...
paths:
...
  /api/orchestrator/advance:
    post:
      tags:
        - orchestrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gateResult: {}
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
  /api/orchestrator/error:
    post:
      tags:
        - orchestrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 2000
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
...
  /api/orchestrator/reset:
    post:
      tags:
        - orchestrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mode
              properties:
                mode:
                  type: string
                  minLength: 1
                  maxLength: 50
                phases:
                  type: array
                  items:
                    type: string
                template:
                  type: string
                  maxLength: 100
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
...
servers:
  - url: http://localhost:3000
    description: Local development server
...